fix(kv): move roster publish to infrastructure for import layers (#1946)
Extract publish_bot_roster and factory-state KV open helper into
infrastructure/kv so agent_cmd and cli_bot no longer transitively import
factory.config via bootstrap. Bootstrap keeps seed-only adapter read path.
