+++ src/factory/infrastructure/kv/__init__.py
@@ -1,0 +1,1 @@
+"""JetStream KV helpers for factory-state."""
